Fullscreen support, event refactoring, and comments
Implemented another game configuration option for fullscreen.

The name for `#trigger_held_buttons` or what it was before was clearly
stating the operation that was being performed. Renamed it through to
the EventRelay.

The interface of `#_draw` was not a good interface and so it has been
renamed to `#draw_with_view`.

Also made the Window's scene an attr_accessor as the scene assignment
was not currently performing any other work.