Scene#scene_name renamed Scenes#create to Scenes#generate
Saw reptitition of scene name calculation so it is now part of scene.

Scenes#create was not clear as it was really #find_and_create so I
renamed it to #generate. However, it may need to be removed or scenes
will need to be able to be created without a window during the
initialize.