Removed Drawers, Promoted Models added SceneActor
To handle the drawing better and the hand-off of data between scenes
with more elegance the SceneActor was born. A SceneActor is a factory
that knows how to create the Actor instance for the scene and takes care
of drawing that actor.

All view items are now simply Actors and not things that they Drawer has
to figure out how to draw.