Model cleanup, comments, and refactoring
Models have their scene set during initialization.

Comments and re-organization of the model class.

Gosu::Color monkeypatching to allow for initialization with existing
colors and strings.

Select now is setup when the window is set and not when the scene is
set. As both pieces are required and the scene usually exists first.