Refactored Events, Drawing, and Animation to modules
The code for storing the animations, events, and drawing appeared more
universal so it is now in modules. This as a bonus obfuscates the
Factories, as in the class names, to the Scene which cares only that it
wants to have these things defined for it.

Refatored to remove a lot of the send calls to use the #actor method to
make the intentions clear what it is that I am asking for in a scene.