Scene#scene_name and Scene#view_name
Scene instances will return the scene name and view name so it is no longer required
to go to the class for that information.

View name will now return the name of the view without the path
information. There is a new class method MyScene.view_path which will
return the path value.