Finalization of Properties Update
Image and Animation use a simplier defaulting syntax.
Image is a simple delegator to a Gosu::Image instead of a monkeypatch.
Models will _load and _save easier now using the properties defined
allowing me to remove the special logic for all that and the workarounds
for saving colors.

Images will default to missing image
FIX: Text incorrectly assumed model was available to be displayed.