Audio Support
Scene.play was added to provide the ability to add a song similar to an
actor being added to draw.

Scene.change and Scene#change are aliases of animate as that is useful
for changing volume levels gradually.

Models will now be told when they are loaded and ready to go similar to
Scenes. This is to allow a point of customization after the window and
scene assets have been added.