Update Gem Dependencies for support of Ruby 2.1 and later
The Gem dependencies were very tightly bound to many Gems that received updates in the past 2 years. It is good practice not to specify version the the gemspec, unless absolute necessary.
This is necessary for texplay. 0.4.4.pre is not only required for Windows development environments, but also for Ruby 2.1 and later on OSX.
The newer versions listen gem don’t offer giving Procs anymore, so this was slightly rewritten.
Gosu::Font has a different constructor now.